requirements/base.txt
Django==1.6.10
bpython==0.13
django-braces==1.4.0
django-model-utils==2.0.3
logutils==0.3.3
South==0.8.4
mosquitto==1.2.3
paho-mqtt==1.0
requests==2.5.3
